refactor Slider tests
