+++ compose/desktop/desktop/src/jvmTest/kotlin/androidx/compose/desktop/InputsTest.kt
@@ -75,256 +75,5 @@ class InputsTest {
         }
         rule.onNodeWithTag(tag).assertRangeInfoEquals(ProgressBarRangeInfo(0f, 0f..1f, 0))
     }
-
-    @OptIn(ExperimentalComposeUiApi::class)
-    @Test
-    fun `slider(0 steps, ltr) changes values when arrows pressed`() {
-        val state = mutableStateOf(0.5f)
-        var sliderFocused = false
-        rule.setContent {
-            Slider(
-                value = state.value,
-                onValueChange = { state.value = it },
-                valueRange = 0f..1f,
-                modifier = Modifier.onFocusChanged {
-                    sliderFocused = it.isFocused
-                }
-            )
-        }
-
-        // Press tab to focus on Slider
-        rule.onRoot().performKeyPress(keyEvent(Key.Tab, KeyEventType.KeyDown))
-        rule.onRoot().performKeyPress(keyEvent(Key.Tab, KeyEventType.KeyUp))
-        rule.runOnIdle {
-            assertTrue(sliderFocused)
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionRight, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionRight, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.50f + (1 + it) / 100f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionLeft, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionLeft, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.53f - (1 + it) / 100f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.PageDown, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.PageDown, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.50f + (1 + it) / 10f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.PageUp, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.PageUp, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.80f - (1 + it) / 10f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionUp, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionUp, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.50f + (1 + it) / 100f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionDown, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionDown, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.53f - (1 + it) / 100f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-
-        rule.onRoot().performKeyPress(keyEvent(Key.MoveEnd, KeyEventType.KeyDown))
-        rule.onRoot().performKeyPress(keyEvent(Key.MoveEnd, KeyEventType.KeyUp))
-        rule.runOnIdle {
-            assertEquals(1f, state.value)
-        }
-
-        rule.onRoot().performKeyPress(keyEvent(Key.Home, KeyEventType.KeyDown))
-        rule.onRoot().performKeyPress(keyEvent(Key.Home, KeyEventType.KeyUp))
-        rule.runOnIdle {
-            assertEquals(0f, state.value)
-        }
-    }
-
-    @OptIn(ExperimentalComposeUiApi::class)
-    @Test
-    fun `slider(0 steps, rtl) changes values when arrows pressed`() {
-        val state = mutableStateOf(0.5f)
-        var sliderFocused = false
-        rule.setContent {
-            CompositionLocalProvider(LocalLayoutDirection provides LayoutDirection.Rtl) {
-                Slider(
-                    value = state.value,
-                    onValueChange = { state.value = it },
-                    valueRange = 0f..1f,
-                    modifier = Modifier.onFocusChanged {
-                        sliderFocused = it.isFocused
-                    }
-                )
-            }
-        }
-
-        // Press tab to focus on Slider
-        rule.onRoot().performKeyPress(keyEvent(Key.Tab, KeyEventType.KeyDown))
-        rule.onRoot().performKeyPress(keyEvent(Key.Tab, KeyEventType.KeyUp))
-        rule.runOnIdle {
-            assertTrue(sliderFocused)
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionRight, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionRight, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.50f - (1 + it) / 100f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionLeft, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionLeft, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((0.47f + (1 + it) / 100f).round2decPlaces(), (state.value).round2decPlaces())
-            }
-        }
-    }
-
-    @OptIn(ExperimentalComposeUiApi::class)
-    @Test
-    fun `slider(29 steps, ltr) changes values when arrows pressed`() {
-        val state = mutableStateOf(15f)
-        var sliderFocused = false
-        rule.setContent {
-            Slider(
-                value = state.value,
-                steps = 29,
-                onValueChange = { state.value = it },
-                valueRange = 0f..30f,
-                modifier = Modifier.onFocusChanged {
-                    sliderFocused = it.isFocused
-                }
-            )
-        }
-
-        // Press tab to focus on Slider
-        rule.onRoot().performKeyPress(keyEvent(Key.Tab, KeyEventType.KeyDown))
-        rule.onRoot().performKeyPress(keyEvent(Key.Tab, KeyEventType.KeyUp))
-        rule.runOnIdle {
-            assertTrue(sliderFocused)
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionRight, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionRight, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((15f + (1f + it)), (state.value))
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionLeft, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionLeft, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((18f - (1 + it)), state.value)
-            }
-        }
-
-        rule.runOnIdle {
-            state.value = 0f
-        }
-
-        val page = ((29 + 1) / 10).coerceIn(1, 10) // same logic as in Slider slideOnKeyEvents
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.PageDown, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.PageDown, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals((1f + it) * page, state.value)
-            }
-        }
-
-        rule.runOnIdle {
-            state.value = 30f
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.PageUp, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.PageUp, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals(30f - (1 + it) * page, state.value)
-            }
-        }
-
-        rule.runOnIdle {
-            state.value = 0f
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionUp, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionUp, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals(1f + it, state.value)
-            }
-        }
-
-        repeat(3) {
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionDown, KeyEventType.KeyDown))
-            rule.onRoot().performKeyPress(keyEvent(Key.DirectionDown, KeyEventType.KeyUp))
-            rule.runOnIdle {
-                assertEquals(3f - (1f + it), state.value)
-            }
-        }
-
-        rule.onRoot().performKeyPress(keyEvent(Key.MoveEnd, KeyEventType.KeyDown))
-        rule.onRoot().performKeyPress(keyEvent(Key.MoveEnd, KeyEventType.KeyUp))
-        rule.runOnIdle {
-            assertEquals(30f, state.value)
-        }
-
-        rule.onRoot().performKeyPress(keyEvent(Key.Home, KeyEventType.KeyDown))
-        rule.onRoot().performKeyPress(keyEvent(Key.Home, KeyEventType.KeyUp))
-        rule.runOnIdle {
-            assertEquals(0f, state.value)
-        }
-    }
 }
 
-private fun Float.round2decPlaces() = (this * 100).roundToInt() / 100f
-
-// TODO: avoid copy-paste. Copy-pasted from KeyInputUtil.kt
-private object DummyComponent : Component()
-/**
- * The [KeyEvent] is usually created by the system. This function creates an instance of
- * [KeyEvent] that can be used in tests.
- */
-fun keyEvent(key: Key, keyEventType: KeyEventType, modifiers: Int = 0): KeyEvent {
-    val action = when (keyEventType) {
-        KeyEventType.KeyDown -> java.awt.event.KeyEvent.KEY_PRESSED
-        KeyEventType.KeyUp -> java.awt.event.KeyEvent.KEY_RELEASED
-        else -> error("Unknown key event type")
-    }
-    return KeyEvent(
-        java.awt.event.KeyEvent(
-            DummyComponent,
-            action,
-            0L,
-            modifiers,
-            key.nativeKeyCode,
-            java.awt.event.KeyEvent.getKeyText(key.nativeKeyCode)[0],
-            key.nativeKeyLocation
-        )
-    )
-}
